Fix: Chat design, security vulnerabilities, and server-wide audit fixes
Chat & WebSocket:
- Prevent message data loss in MessageBuffer (clear after save, re-queue on failure)
- Add ORDER BY to message query (was loading in random order)
- Persist messageStatus and isRead on Message entity
- Update Chat.modifiedAt when new messages are saved
- Add findByUserId query to replace findAll+filter in ChatService
- Client: add id field to ChatMessage, improve dedup logic

Security:
- Fix path traversal in FileStorageServiceImpl (UUID filename + path check)
- Constant-time secret key comparison (prevent timing attacks)
- Fix JSON injection in filter error responses
- Remove filePath from AttachmentResponse (leaked server paths)
- Reduce BCrypt strength 16→12 (was 3-5s per hash, DoS risk)
- Add password max length (128) to prevent BCrypt DoS

Data integrity:
- Fix ApiKeyRepository Short→Long type mismatch (runtime crash)
- Fix TokenService.refreshToken null expiration (tokens never expired)
- Add null guard on ApiKeyService.validateApiKey expiredAt
- Add unique+nullable constraints on Token.token, ApiKey.apiKey, User.password
- Add DB indexes on Message(chatId,userId) and Chat(senderId,recipientId)
- Fix UserMapper: ignore password mapping, set isVerified default

Validation & error handling:
- Add Bean Validation to 5 DTOs
- Add MethodArgumentNotValidException handler for consistent 400 responses
- Log stack traces in GlobalExceptionHandler catch-all
- Add null guards on PlacesService RestClient responses
- Remove duplicate validation dependency from pom.xml
- Add explicit @Mapping ignores on ChatMapper and UserMapper

+++ server/src/main/java/com/ump/studentgrab/application/dto/response/ApiKeyResponse.java
@@ -3,7 +3,7 @@ package com.ump.studentgrab.application.dto.response;
 import java.time.LocalDateTime;
 
 public record ApiKeyResponse(
-        Short id,
+        Long id,
         String apiKey,
         Long createdBy,
         LocalDateTime createdAt,
