+++ server/src/main/java/com/ump/studentgrab/domain/repository/ApiKeyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ApiKeyRepository extends JpaRepository<ApiKey, Short> {
+public interface ApiKeyRepository extends JpaRepository<ApiKey, Long> {
 
     Optional<ApiKey> findByApiKey(String apiKey);
 }
